// thirdeye-ui/src/app/components/entity-cards/subscription-group-view-card/subscription-group-view-card.interfaces.ts
/*
 * Copyright 2023 StarTree Inc
 *
 * Licensed under the StarTree Community License (the "License"); you may not use
 * this file except in compliance with the License. You may obtain a copy of the
 * License at http://www.startree.ai/legal/startree-community-license
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the
 * License is distributed on an "AS IS" BASIS, WITHOUT * WARRANTIES OF ANY KIND,
 * either express or implied.
 *
 * See the License for the specific language governing permissions and limitations under
 * the License.
 */

import { ReactNode } from "react";

export interface SubscriptionGroupViewCardProps {
    header?: string | ReactNode;
    rows: { label: string; value: string | ReactNode }[];
}